wip: use global redux state to manage opened map tool
