+++ frontend/src/features/map/tools/interest_points/InterestPointMapButton.js
@@ -1,32 +1,31 @@
 import React, { useEffect, useMemo, useRef } from 'react'
 import styled from 'styled-components'
 import { useDispatch, useSelector } from 'react-redux'
-import { COLORS } from '../../constants/constants'
-import { expandRightMenu, setMapToolOpened } from '../../domain/shared_slices/Global'
-import { MapButtonStyle } from '../commonStyles/MapButton.style'
-import { ReactComponent as InterestPointSVG } from '../icons/Point_interet.svg'
+import { COLORS } from '../../../../constants/constants'
+import { expandRightMenu, setMapToolOpened } from '../../../../domain/shared_slices/Global'
+import { MapButtonStyle } from '../../../commonStyles/MapButton.style'
+import { ReactComponent as InterestPointSVG } from '../../../icons/Point_interet.svg'
 import EditInterestPoint from './EditInterestPoint'
 import {
   deleteInterestPointBeingDrawed,
   drawInterestPoint,
   endInterestPointDraw
-} from '../../domain/shared_slices/InterestPoint'
-import { useEscapeFromKeyboard } from '../../hooks/useEscapeFromKeyboard'
-import { MapTool } from '../../domain/entities/map'
+} from '../../../../domain/shared_slices/InterestPoint'
+import { useEscapeFromKeyboard } from '../../../../hooks/useEscapeFromKeyboard'
+import { MapTool } from '../../../../domain/entities/map'
+import { MapToolButton } from '../MapToolButton'
 
-const InterestPoint = () => {
+const InterestPointMapButton = () => {
   const dispatch = useDispatch()
-  const {
-    isEditing
-  } = useSelector(state => state.interestPoint)
   const {
     healthcheckTextWarning,
     rightMenuIsOpen,
     previewFilteredVesselsMode,
     mapToolOpened
   } = useSelector(state => state.global)
+  const selectedVessel = useSelector(state => state.vessel.selectedVessel)
 
-  const isRightMenuShrinked = !rightMenuIsOpen
+  const isRightMenuIconShrinked = selectedVessel && !rightMenuIsOpen
   const isInterestPointOpen = useMemo(() => mapToolOpened === MapTool.INTEREST_POINT, [mapToolOpened])
   const wrapperRef = useRef(null)
   const escapeFromKeyboard = useEscapeFromKeyboard()
@@ -39,10 +38,8 @@ const InterestPoint = () => {
 
   function openOrCloseInterestPoint () {
     if (!isInterestPointOpen) {
-      if (!isEditing) {
-        dispatch(drawInterestPoint())
-        dispatch(setMapToolOpened(MapTool.INTEREST_POINT))
-      }
+      dispatch(drawInterestPoint())
+      dispatch(setMapToolOpened(MapTool.INTEREST_POINT))
     } else {
       close()
     }
@@ -51,27 +48,25 @@ const InterestPoint = () => {
   function close () {
     dispatch(endInterestPointDraw())
     dispatch(setMapToolOpened(undefined))
-    if (!isEditing) {
-      dispatch(deleteInterestPointBeingDrawed())
-    }
+    dispatch(deleteInterestPointBeingDrawed())
   }
 
   return (
     <Wrapper ref={wrapperRef}>
-      <InterestPointWrapper
+      <InterestPointButton
         data-cy={'interest-point'}
         isHidden={previewFilteredVesselsMode}
         healthcheckTextWarning={healthcheckTextWarning}
         isOpen={isInterestPointOpen}
-        rightMenuIsShrinked={isRightMenuShrinked}
+        rightMenuIsShrinked={!rightMenuIsOpen}
         onMouseEnter={() => dispatch(expandRightMenu())}
         title={'Créer un point d\'intérêt'}
         onClick={openOrCloseInterestPoint}
       >
         <InterestPointIcon
-          $rightMenuIsShrinked={isRightMenuShrinked}
+          $rightMenuIsShrinked={!rightMenuIsOpen}
         />
-      </InterestPointWrapper>
+      </InterestPointButton>
       <EditInterestPoint
         healthcheckTextWarning={healthcheckTextWarning}
         isOpen={isInterestPointOpen}
@@ -86,25 +81,8 @@ const Wrapper = styled.div`
   z-index: 1000;
 `
 
-const InterestPointWrapper = styled(MapButtonStyle)`
-  position: absolute;
-  display: inline-block;
-  background: ${props => props.isOpen ? COLORS.shadowBlue : COLORS.charcoal};
+const InterestPointButton = styled(MapToolButton)`
   top: 291px;
-  z-index: 99;
-  height: 40px;
-  width: ${props => props.rightMenuIsShrinked ? '5px' : '40px'};
-  border-radius: ${props => props.rightMenuIsShrinked ? '1px' : '2px'};
-  right: ${props => props.rightMenuIsShrinked ? '0' : '10px'};
-  transition: all 0.3s;
-
-  :hover {
-      background: ${COLORS.charcoal};
-  }
-
-  :focus {
-      background: ${COLORS.shadowBlue};
-  }
 `
 
 const InterestPointIcon = styled(InterestPointSVG)`
@@ -113,4 +91,4 @@ const InterestPointIcon = styled(InterestPointSVG)`
   transition: all 0.2s;
 `
 
-export default InterestPoint
+export default InterestPointMapButton
