+++ frontend/src/features/map/tools/vessel_filters/Filter.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react'
 import styled from 'styled-components'
-import { COLORS } from '../../constants/constants'
-import { ReactComponent as CloseIconSVG } from '../icons/Croix_grise.svg'
-import { ReactComponent as ShowIconSVG } from '../icons/oeil_affiche.svg'
-import { ReactComponent as HideIconSVG } from '../icons/oeil_masque.svg'
-import { ReactComponent as ChevronIconSVG } from '../icons/Chevron_simple_gris.svg'
-import { ReactComponent as FilterSVG } from '../icons/Icone_filtres_dark.svg'
+import { COLORS } from '../../../../constants/constants'
+import { ReactComponent as CloseIconSVG } from '../../../icons/Croix_grise.svg'
+import { ReactComponent as ShowIconSVG } from '../../../icons/oeil_affiche.svg'
+import { ReactComponent as HideIconSVG } from '../../../icons/oeil_masque.svg'
+import { ReactComponent as ChevronIconSVG } from '../../../icons/Chevron_simple_gris.svg'
+import { ReactComponent as FilterSVG } from '../../../icons/Icone_filtres_dark.svg'
 import TagList from './TagList'
 
 const Filter = ({ filter, index, isLastItem, removeFilter, showFilter, hideFilters, removeTagFromFilter }) => {
@@ -117,7 +117,7 @@ const ChevronIcon = styled(ChevronIconSVG)`
   width: 16px;
   margin-right: 8px;
   margin-top: 5px;
-  
+
   animation: ${props => props.$isOpen ? 'chevron-layer-opening' : 'chevron-layer-closing'} 0.5s ease forwards;
 
   @keyframes chevron-layer-opening {
