+++ frontend/src/features/map/tools/vessel_filters/FilterParameters.js
@@ -2,12 +2,12 @@ import React from 'react'
 import styled from 'styled-components'
 import { useDispatch } from 'react-redux'
 
-import { openVesselListModal } from '../../domain/shared_slices/Global'
-import MapPropertyTrigger from '../commonComponents/MapPropertyTrigger'
+import { openVesselListModal } from '../../../../domain/shared_slices/Global'
+import MapPropertyTrigger from '../../../commonComponents/MapPropertyTrigger'
 
-import { ReactComponent as HidingOtherVesselsSVG } from '../icons/Bouton_masquer_pistes_actif.svg'
-import { ReactComponent as ShowingOtherVesselsSVG } from '../icons/Bouton_masquer_pistes_inactif.svg'
-import { COLORS } from '../../constants/constants'
+import { ReactComponent as HidingOtherVesselsSVG } from '../../../icons/Bouton_masquer_pistes_actif.svg'
+import { ReactComponent as ShowingOtherVesselsSVG } from '../../../icons/Bouton_masquer_pistes_inactif.svg'
+import { COLORS } from '../../../../constants/constants'
 
 const FilterParameters = ({ nonFilteredVesselsAreHidden, setNonFilteredVesselsAreHidden }) => {
   const dispatch = useDispatch()
